requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anthropic==0.37.1
anyio==4.6.2.post1
attrs==24.2.0
black==24.10.0
blinker==1.8.2
cattrs==23.2.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dill==0.3.9
distro==1.9.0
dnspython==2.7.0
ell-ai==0.0.14
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.5
filelock==3.16.1
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy-extensions==1.0.0
numpy==2.1.2
openai==1.53.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
psutil==5.9.8
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.16
PyYAML==6.0.2
requests==2.32.3
rich==13.9.3
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.21
starlette==0.37.2
tokenizers==0.20.1
tqdm==4.66.6
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
Werkzeug==3.0.6
